refactor: getMovieDetail 메서드를 domain으로 분리
